langGraph-base-py/requirements.txt
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
langchain-tavily>=0.1.0
langgraph>=0.3.0
python-dotenv>=1.0.0 